+++ tests/testthat/test-tf-idf.R
@@ -45,3 +45,35 @@ test_that("TF-IDF works when the document ID is a number", {
   expect_false(any(is.na(tf_idf)))
   expect_equal(tf_idf$tf_idf[c(3, 6)], c(0, 0))
 })
+
+
+test_that("tf-idf with tidyeval works", {
+  d <- data_frame(txt = c("Because I could not stop for Death -",
+                          "He kindly stopped for me -"))
+  termvar <- quo("word")
+  documentvar <- quo("document")
+  countvar <- quo("frequency")
+
+  result <- w %>%
+    bind_tf_idf(!! termvar, !! documentvar, !! countvar)
+
+  expect_equal(select(w, document, word, frequency),
+               select(result, document, word, frequency))
+
+  expect_is(result, "tbl_df")
+  expect_is(result$tf, "numeric")
+  expect_is(result$idf, "numeric")
+  expect_is(result$tf_idf, "numeric")
+
+  expect_equal(result$tf, rep(c(1 / 6, 1 / 6, 1 / 6, 1 / 6, 1 / 3), 2))
+  expect_equal(result$idf[1:4], c(0, log(2), 0, log(2)))
+  expect_equal(result$tf_idf, result$tf * result$idf)
+
+  result2 <- w %>%
+    group_by(document) %>%
+    bind_tf_idf(!! termvar, !! documentvar, !! countvar)
+
+  #expect_equal(result, ungroup(result2))
+  expect_equal(length(groups(result2)), 1)
+  expect_equal(as.character(groups(result2)[[1]]), "document")
+})