+++ NEWS.md
@@ -2,7 +2,7 @@
 
 * Fix tidier for quanteda dictionary for correct class (#71).
 * Add a [pkgdown site](https://juliasilge.github.io/tidytext).
-* Convert NSE from underscored function to tidyeval (`unnest_tokens`) (#67, #74).
+* Convert NSE from underscored function to tidyeval (`unnest_tokens`, `bind_tf_idf`) (#67, #74).
 * Added tidiers for topic models from the `stm` package (#51).
 
 # tidytext 0.1.3